requirements.txt
Flask
flask-cors
python-dotenv
openai
whisper
bcrypt
pyjwt